Remove 1 tick delay from piston
Imagine there is a lever connected to a piston via some wire. It takes 1 tick for the lever to become powered. It takes 1 tick for the wire to become powered. It takes 1 tick for the piston head to arrive at its new location. Previously, there was 1 tick delay before the piston because powered and then the piston would start moving after that.