Add power sprite writer
Thats the rendering system done! Tomorrow, I'm going to define all of the entities and test all of code. Come to think of it, maybe I should have started defining the entities a little earlier so that I could test the code as I was writing it (instead of testing it on a whiteboard). Well, its too late now!