Create assemble function
An assembly is a function that generates an array of entities. Assemblies can be arbitrarily nested. The assemble function recursively flattens the assemblies.

A good example of an assembly is a rectangle. A rectangle can take the position, size and entity as parameters and produce an array of entities that form a rectangle.