Add Delay component
Adding a new entity this late might not be a good idea. I was planning on having an entity with semantics similar to a Minecraft redstone repeater but I couldn't figure out to implement it. I didn't want to use a queue. I wanted the entity to extend the pulse of the input. I think I've figured out how to implement it so I'm going to create the delay entity.