Create delay entity
It seems to work exactly as I wanted it to. I'll have to draw some sprites for it though. I'm using the identity gate for now.

I've noticed that the NOR clock is 2/4 ticks. I want to find a way to create a 1/2 clock. I also need a way to create a rising and falling edge detector that outputs 1 tick.