Allow the animation speed of a static sprite to be slowed
There's no need to make an animation faster because you could just draw less frames and then repeat them. For example, to double to speed of an animation, draw 5 frames and then repeat the 5 frames so that you have 10 frames. Slowing down an animation is possible without some help from the engine