Animate start sequence of title screen
The title screen has two animations. The first animation plays once when the window is first opened. The second animation plays after the first and loops until the player presses a key. The second animation has the message "Press any key to start" on it so that the player knows what to do. I've created the first animation.