Create Levels folder
Entities, Assemblies and Levels are all in the same folder to make imports a little shorter. I did this so that I could specify input and output files to the level compiler script. This speeds up launch because levels are only compiled if they need to be. The problem is that levels ALL levels are compiled if ONE of the levels changes. There isn't really an element way around this so I guess this is the best I can do.