Create level 2
I've changed the build script so that all files are created even if one of them fails to compile. This means that only one output file has to be specified.

Xcode will run the script if any of the input files have changed or any of the output files are not present. The script always creates all output files so all output files are either present or not. This means that we only need to specify one output file.