Create level 6
I'm hoping that players don't immediately notice that the exit is accessible. By level 6 they should have it drilled into their heads that they must first go up, do something, then go back down and to the exit. It's impossible to get the box to the pressure plate so the player should struggle with that for a few seconds before going to the exit.